% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart.R
\name{chart}
\alias{chart}
\title{Shoe size Charts}
\usage{
chart(gender = "female")
}
\arguments{
\item{gender}{The gender of the user, either male or female, default to female}
}
\value{
A shoe size chart
}
\description{
This function allows the user to print out a shoe size conversion chart.
}
\details{
If you think the function is too good to be true, check it out for youself! Source from \url{https://www.zappos.com/c/shoe-size-conversion}
}
\examples{
chart('male')
}
\author{
Chuze Zhang <\url{https://github.com/QMSS-G5072-2019/Zhang_Chuze}>
}
