% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shoesize.R
\name{converter}
\alias{converter}
\title{Helper Function}
\usage{
converter(size_input, size_standard, ussize_standard)
}
\arguments{
\item{size_input}{User's UK or EU shoe size}

\item{size_standard}{UK or EU standard size}

\item{ussize_standard}{US standard size}
}
\value{
User's US shoe size
}
\description{
This function is a helper function that helps the main function to convert shoe sizes.
}
\author{
Chuze Zhang <\url{https://github.com/QMSS-G5072-2019/Zhang_Chuze}>
}
\keyword{internal}
