% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_shoe.R
\name{search_shoe}
\alias{search_shoe}
\title{Online Shopping Search}
\usage{
search_shoe(gender = "female", brand = "")
}
\arguments{
\item{gender}{The gender of the user, either male or female, default to female}

\item{brand}{Input your favorite shoe brand}
}
\value{
Online search result
}
\description{
This function allows the user to search for his or her favorite shoe brand online.
}
\details{
Now you have your US shoe size, let's go shopping!
}
\examples{
search_shoe('male','Nike')
}
\author{
Chuze Zhang <\url{https://github.com/QMSS-G5072-2019/Zhang_Chuze}>
}
