% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shoesize.R
\name{shoesize}
\alias{shoesize}
\title{Shoe Size Converter}
\usage{
shoesize(gender = "female", country = "uk", size = 0)
}
\arguments{
\item{gender}{The gender of the user, either male or female, default to female}

\item{country}{Which sizing standard does the user use, UK or Europe, default to UK}

\item{size}{User's UK or EU shoe size}
}
\value{
User's US shoe size
}
\description{
This function is the main function in this package. It allows users to convert their UK or EU shoe sizes to US sizes.
}
\examples{
shoesize('male', 'Euro', 42)
shoesize('Female', 'UK', 5)
}
\author{
Chuze Zhang <\url{https://github.com/QMSS-G5072-2019/Zhang_Chuze}>
}
